Consolidate ping target setting into settings.py
